pyproject.toml
[tool.poetry]
name = "manim_speech"
version = "0.3.0-dev"
description = "Manim plugin for adding speech to videos."
authors = ["Celeste Ma <mtcelestema@proton.me>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/MtCelesteMa/manim-speech"
repository = "https://github.com/MtCelesteMa/manim-speech"
keywords = ["manim", "speech", "voiceover", "tts"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12"
]
packages = [
    {include = "manim_speech", from = "src"}
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
manim = "^0.18.1"
pydantic = "^2.8.0"
mutagen = "^1.47.0"
polib = "^1.2.0"
python-slugify = "^8.0.4"
openai = {version = "^1.35.9", optional = true}
elevenlabs = {version = "^1.3.1", optional = true}
openai-whisper = {git = "https://github.com/openai/whisper.git", rev = "8bc8860694949db53c42ba47ddc23786c2e02a8b", optional = true}
assemblyai = {version = "^0.28.1", optional = true}
deepl = {version = "^1.18.0", optional = true}
langconv = {version = "^0.3.0", optional = true}

[tool.poetry.extras]
openai = ["openai"]
elevenlabs = ["elevenlabs"]
whisper = ["openai-whisper"]
assemblyai = ["assemblyai"]
deepl = ["deepl"]
chinese = ["langconv"]

[tool.poetry.group.dev.dependencies]
tox = "^4.16.0"
isort = "^5.13.2"
ruff = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
